+++ packages/liveblocks-core/src/comments/index.ts
@@ -1,9 +1,18 @@
 import type { AuthValue } from "../auth-manager";
 import type { JsonObject } from "../lib/Json";
+import {
+  convertToCommentData,
+  convertToCommentUserReaction,
+  convertToThreadData,
+} from "./comment-body";
 import type { BaseMetadata } from "./types/BaseMetadata";
 import type { CommentBody } from "./types/CommentBody";
-import type { CommentData } from "./types/CommentData";
-import type { ThreadData } from "./types/ThreadData";
+import type { CommentData, CommentDataPlain } from "./types/CommentData";
+import type {
+  CommentUserReaction,
+  CommentUserReactionPlain,
+} from "./types/CommentReaction";
+import type { ThreadData, ThreadDataPlain } from "./types/ThreadData";
 
 type Options = {
   baseUrl: string;
@@ -21,8 +30,20 @@ type PartialNullable<T> = {
   [P in keyof T]?: T[P] | null | undefined;
 };
 
+export type QueryParams =
+  | Record<string, string | number | boolean | null | undefined>
+  | URLSearchParams;
+
+export type ThreadsFilterOptions<TThreadMetadata extends BaseMetadata> = {
+  query?: {
+    metadata?: Partial<TThreadMetadata>;
+  };
+};
+
 export type CommentsApi<TThreadMetadata extends BaseMetadata> = {
-  getThreads(): Promise<ThreadData<TThreadMetadata>[]>;
+  getThreads(
+    options?: ThreadsFilterOptions<TThreadMetadata>
+  ): Promise<ThreadData<TThreadMetadata>[]>;
   createThread(options: {
     threadId: string;
     commentId: string;
@@ -32,7 +53,7 @@ export type CommentsApi<TThreadMetadata extends BaseMetadata> = {
   editThreadMetadata(options: {
     metadata: PartialNullable<TThreadMetadata>;
     threadId: string;
-  }): Promise<ThreadData<TThreadMetadata>>;
+  }): Promise<TThreadMetadata>;
   createComment(options: {
     threadId: string;
     commentId: string;
@@ -51,12 +72,12 @@ export type CommentsApi<TThreadMetadata extends BaseMetadata> = {
     threadId: string;
     commentId: string;
     emoji: string;
-  }): Promise<CommentData>;
+  }): Promise<CommentUserReaction>;
   removeReaction(options: {
     threadId: string;
     commentId: string;
     emoji: string;
-  }): Promise<CommentData>;
+  }): Promise<void>;
 };
 
 export class CommentsApiError extends Error {
@@ -121,8 +142,9 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     const url = new URL(
       `/v2/c/rooms/${encodeURIComponent(roomId)}${endpoint}`,
       config.baseUrl
-    ).toString();
-    return await fetch(url, {
+    );
+
+    return await fetch(url.toString(), {
       ...options,
       headers: {
         ...options?.headers,
@@ -131,14 +153,24 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     });
   }
 
-  async function getThreads(): Promise<ThreadData<TThreadMetadata>[]> {
-    const response = await fetchApi(roomId, "/threads");
+  async function getThreads(
+    options?: ThreadsFilterOptions<TThreadMetadata>
+  ): Promise<ThreadData<TThreadMetadata>[]> {
+    const response = await fetchApi(roomId, "/threads/search", {
+      body: JSON.stringify({
+        ...(options?.query?.metadata && { metadata: options.query.metadata }),
+      }),
+      headers: {
+        "Content-Type": "application/json",
+      },
+      method: "POST",
+    });
 
     if (response.ok) {
       const json = await (response.json() as Promise<{
-        data: ThreadData<TThreadMetadata>[];
+        data: ThreadDataPlain<TThreadMetadata>[];
       }>);
-      return json.data;
+      return json.data.map((thread) => convertToThreadData(thread));
     } else if (response.status === 404) {
       return [];
     } else {
@@ -146,7 +178,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     }
   }
 
-  function createThread({
+  async function createThread({
     metadata,
     body,
     commentId,
@@ -158,23 +190,28 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     metadata: TThreadMetadata | undefined;
     body: CommentBody;
   }) {
-    return fetchJson<ThreadData<TThreadMetadata>>("/threads", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        id: threadId,
-        comment: {
-          id: commentId,
-          body,
+    const thread = await fetchJson<ThreadDataPlain<TThreadMetadata>>(
+      "/threads",
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
         },
-        metadata,
-      }),
-    });
+        body: JSON.stringify({
+          id: threadId,
+          comment: {
+            id: commentId,
+            body,
+          },
+          metadata,
+        }),
+      }
+    );
+
+    return convertToThreadData(thread);
   }
 
-  function editThreadMetadata({
+  async function editThreadMetadata({
     metadata,
     threadId,
   }: {
@@ -182,7 +219,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     metadata: PartialNullable<TThreadMetadata>;
     threadId: string;
   }) {
-    return fetchJson<ThreadData<TThreadMetadata>>(
+    return await fetchJson<TThreadMetadata>(
       `/threads/${encodeURIComponent(threadId)}/metadata`,
       {
         method: "POST",
@@ -194,7 +231,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     );
   }
 
-  function createComment({
+  async function createComment({
     threadId,
     commentId,
     body,
@@ -203,7 +240,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     body: CommentBody;
   }) {
-    return fetchJson<CommentData>(
+    const comment = await fetchJson<CommentDataPlain>(
       `/threads/${encodeURIComponent(threadId)}/comments`,
       {
         method: "POST",
@@ -216,9 +253,11 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
         }),
       }
     );
+
+    return convertToCommentData(comment);
   }
 
-  function editComment({
+  async function editComment({
     threadId,
     commentId,
     body,
@@ -227,7 +266,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     body: CommentBody;
   }) {
-    return fetchJson<CommentData>(
+    const comment = await fetchJson<CommentDataPlain>(
       `/threads/${encodeURIComponent(threadId)}/comments/${encodeURIComponent(
         commentId
       )}`,
@@ -241,6 +280,8 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
         }),
       }
     );
+
+    return convertToCommentData(comment);
   }
 
   async function deleteComment({
@@ -261,7 +302,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     );
   }
 
-  function addReaction({
+  async function addReaction({
     threadId,
     commentId,
     emoji,
@@ -270,7 +311,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     emoji: string;
   }) {
-    return fetchJson<CommentData>(
+    const reaction = await fetchJson<CommentUserReactionPlain>(
       `/threads/${encodeURIComponent(threadId)}/comments/${encodeURIComponent(
         commentId
       )}/reactions`,
@@ -282,9 +323,11 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
         body: JSON.stringify({ emoji }),
       }
     );
+
+    return convertToCommentUserReaction(reaction);
   }
 
-  function removeReaction({
+  async function removeReaction({
     threadId,
     commentId,
     emoji,
@@ -293,7 +336,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     emoji: string;
   }) {
-    return fetchJson<CommentData>(
+    await fetchJson<CommentData>(
       `/threads/${encodeURIComponent(threadId)}/comments/${encodeURIComponent(
         commentId
       )}/reactions/${encodeURIComponent(emoji)}`,