+++ packages/liveblocks-core/src/comments/types/CommentReaction.ts
@@ -1,0 +1,9 @@
+import type { DateToString } from "./DateToString";
+
+export type CommentUserReaction = {
+  emoji: string;
+  createdAt: Date;
+  userId: string;
+};
+
+export type CommentUserReactionPlain = DateToString<CommentUserReaction>;